Move Data tracking to BridgeApp in order to simplify tracking with our services.
