+++ ResearchStack2/ResearchStack2/RSDDocumentable.swift
@@ -118,9 +118,6 @@ public struct RSDDocumentCreator {
             RSDDurationRangeObject.self,
             RSDFetchableImageThemeElementObject.self,
             RSDFileResultObject.self,
-            RSDMedicationAnswer.self,
-            RSDMedicationItem.self,
-            RSDMedicationTrackingResult.self,
             RSDNumberRangeObject.self,
             RSDResourceTransformerObject.self,
             RSDResultObject.self,
@@ -128,9 +125,6 @@ public struct RSDDocumentCreator {
             RSDStandardAsyncActionConfiguration.self,
             RSDTaskResultObject.self,
             RSDTextFieldOptionsObject.self,
-            RSDTrackedItemObject.self,
-            RSDTrackedItemsResultObject.self,
-            RSDTrackedSectionObject.self,
             RSDUIActionObject.self,
             RSDViewThemeElementObject.self,
             RSDWebViewUIActionObject.self,
@@ -165,8 +159,6 @@ public struct RSDDocumentCreator {
         RSDFormUIStepObject.self,
         RSDOverviewStepObject.self,
         RSDSectionStepObject.self,
-        RSDTrackedItemsReviewStepObject.self,
-        RSDTrackedSelectionStepObject.self,
         RSDStepTransformerObject.self,
         RSDInputFieldObject.self,
         RSDChoiceInputFieldObject.self,
