+++ ResearchStack2/ResearchStack2/RSDStepType.swift
@@ -62,20 +62,11 @@ public struct RSDStepType : RawRepresentable, Codable {
     /// Defaults to creating a `RSDActiveUIStep`.
     public static let instruction: RSDStepType = "instruction"
     
-    /// Defaults to creating a `RSDTrackedItemsLoggingStepObject`.
-    public static let logging: RSDStepType = "logging"
-    
     /// Defaults to creating a `RSDOverviewStepObject`.
     public static let overview: RSDStepType = "overview"
-    
-    /// Defaults to creating a `RSDTrackedItemsReviewStepObject`.
-    public static let review: RSDStepType = "review"
-    
+
     /// Defaults to creating a `RSDSectionStep`.
     public static let section: RSDStepType = "section"
-    
-    /// Defaults to creating a `RSDTrackedSelectionStepObject`.
-    public static let selection: RSDStepType = "selection"
     
     /// Defaults to creating a `RSDSectionStep` created using a `RSDTransformerStep`.
     public static let transform: RSDStepType = "transform"
